fix: handling of config options (#85)
This PR adds 2 configuration options and fixes handling of existing
ones:
- `fund_dev_accounts` option is added to `network_params`, allowing to
configure whether dev accounts should be funded on L2. In one of recent
updates they stopped being funded by default, this PR changes it, let me
know if this is undesired
- adds `batcher_image` option allowing to configure custom op-batcher
image
- fixes handling of `el_extra_params`, right now this argument is not
respected
- fixes handling of `seconds_per_slot`, right now this argument is not
respected

Also it seems that `count` key for participants is no more respected. Is
this intentional?