fix: don't truncate names too arbitrarily
The problem with selecting the last 63 characters of the image name
for a k8s label is that there's no guarantee that the result will
respect the other contraints.

In particular,
us-docker.pkg.dev/oplabs-tools-artifacts/images/op-geth:v1.101411.4-rc.4
yields .pkg.dev/oplabs-tools-artifacts/images/op-geth:v1.101411.4-rc.4
which is invalid because it starts with a dot.
So the net outcome is that valid images fail to deploy because of an
invalid attached label.

This change select the longest prefix available while breaking only at /
So for the image above, the label would be
oplabs-tools-artifacts/images/op-geth:v1.101411.4-rc.4