feat(test): Unit testing using kurtestosis (#161)
**Description**

[`kurtestosis`](https://github.com/ethereum-optimism/kurtestosis),
currently an MVP, is a unit test runner for `starlark`/`kurtosis`. It
allows for lower-level testing that would otherwise require costly and
complex environment setup and introduce unjustifiable test code
complexity.

This PR adds a PoC test for a simple piece of functionality,
`label_from_image` function in `util.star`. The test is then executed as
a part of the PR pipeline. `mise` is introduced to manage dependencies.