chore: add a challenger 'enabled' flag to make its deployment optional (#165)
challenger and its dependency (building the op-program cannon preimage)
take a long time to build in op's monorepo.

This change is a first step in being able to disable the challenger when
not needed. Will need to figure out a way to also make the cannon
preimage building optional from op's repo.