feat(op-deployer): optionally receive contracts from enclave artifact (#203)
This change enables referencing contracts that are present in a local
artifact, in addition to the existing ability to use a remote URL.

In particular, it makes it easier to use locally-built contracts without
having to maintain a separate webserver just to serve them (see
https://github.com/ethereum-optimism/optimism/issues/14390)

More generally, downstream consumers of this package can make use of
the capability to keep relevant artifacts within the enclave.

This works by interpreting a pseudo-URL of the form artifact://NAME
and mapping the corresponding artifact onto a file:// URL for
op-deployer's benefit.