fix(util): name artifacts created by write_to_file (#205)
Calls to write_file are currently the only ones that create artifacts
with generated names, which is needlessly getting in the way of
looking at the outcomes.

We have 2 call sites for this function:
- dependency_set.json
- intent.json

Arguably the latter is not super useful to expose, as it's an
intermediate representation of what eventually lands in
op-deployer-configs, but it doesn't hurt either.