feat: add op-faucet component (#228)
This adds preliminary support for op-faucet.

see https://github.com/ethereum-optimism/optimism/tree/develop/op-faucet

An official image doesn't exist for it yet, so there will need to be a
followup once this is addressed.

This is a test-only construct.
When enabled, we have an unrestricted faucet tied to the (funded)
faucet wallets.

Each chain in the devnet (l1, and all l2s) are mapped to this faucet.