chore: Support multiple supervisors: utilities [1/N] (#237)
**Description**

The multiple supervisor support will require a rework of a significant
part of the challenger, L2 and supervisor code. To make the new code
DRYer, we are creating utilities for working with dicts and lists:

- `filter.star`
- `remove_none` for removing `None` values from `dict`s. YAML supports
`None` o it's important we strip these out before merging any
dictionaries with default values
- `remove_keys` for removing a list of keys from `dict`s. This will be a
part of the replacement for the centralized `sanity_check` functions -
we will use it to fail the process if we find any unknown properties on
a configuration object.
- `expansion.star`
- `expand_asterisc` for working with special `"*"` values that mean
_all_. This will be used in places where a configuration object needs to
point to multiple objects, e.g. lists of L2s for challenger

The utilities have been placed in their individual files as opposed to
the existing `util.star` that has gotten a bit out of control and mixes
functions of all sorts into one big file.

Related to https://github.com/ethereum-optimism/optimism/issues/15611