chore: Support multiple supervisors: Decouple challengers from networks [4/N] (#243)
**Description**

The challenger instances are decoupled from L2 networks and their
configuration is moving to the top-level configuration.

**Notes**

There are several larger motions in this PR aimed at laying down the
groundwork for further improvements:

- The intermediate step of parsing the input to `dict`, then
reformatting it to a `struct` has been removed. As it is, it adds no
value and only creates room for error
- The input parser for challenger has been moved to its own file
- Challenger configuration is structured as a dictionary in YAML - this
enforces unique and YAML-valid (no whitespace, etc) names, which removes
the need for extra validation logic. The dictionary is then parsed into
a list for ease of enumeration, and the names are added to the config
objects themselves
- There is no _default configuration_ for challenger (in the sense that
there is no `challengers.default_params` object. Instead, YAML anchors
can be utilised to share pieces of configuration. This removes the need
for extra parsing logic and keeps things as lean as possible
- The config objects are enriched with `service_name` property. This
will become important when it comes to resolving circular dependencies
later on, as URLs will be possible to construct without having an
instance of `Service`
- The ubiquitous `mkdir &&` combined with `sh -c` entrypoint has been
replaced (only for `op-challenger`) with a persistent directory.

**Questions**

@sigma there will be more objects where something points to a list of
networks. I named the field `participants` so that later on we might
expand it to cover both networks and interop sets (with a heterogeneous
list of some sort, TBD). That okay? I want to avoid confusion with
`participants` in `chains` that signify nodes rather than networks.

Related to https://github.com/ethereum-optimism/optimism/issues/15611