fix(op-deployer): allow overriding cannon vm type (#249)
This adds the ability to override the vmType as part of op-deployer
operations.

main use-case will probably be to use CANNON-NEXT in some experiments.

We also rename global_deployer_overrides to simply overrides as it
doesn't properly match op-deployer's intent's globalDeployOverrides
concept anymore.

Also rework the overrides mechanism so we can actually specify a subset
of the overrides.