feat: add overridable registry (#253)
This change adds a concept of registry for the various container images
we handle.

It has several purposes:
- to centralize the definition of the container images we use
- to make it overridable by a downstream consumer

In particular, this improves the reprodicibility capabilities of the
package, as said downstream consumers can pin specific versions of the
packages being used, in a way that's orthogonal to the arguments
definition that's passed to the package.

See https://github.com/ethereum-optimism/optimism/pull/15697 for an
example.
The *defaults* are pinned, without affecting the ability to override
images on a per-component instance basis.