chore: Support multiple sequencers: Stub of L2 network participant input parsers [1/N] (#268)
**Description**

- Adds new input parsers for EL, CL and adds them under a participant
parser. The prefixed version (`el_something_something`,
`c_something_something`) has been replaced by nested objects (prefixes
typically point to a lack of nesting), which simplifies the parsing
logic
- Despite the EL & CL parsing being quite similar, I kept them in
separate parsers to allow them to diverge and not to obscure simple
parsing with complex abstractions. We'll see if this stands the test of
time
- `labels` have been added to the parsed params to see whether we could
use those

The names of the services have changed and should be more readable
(which will incur a cost on the devnet SDK parsing side). I am not set
on this particular format, but improvement is necessary, feel free to
suggest different format.

The participant config does not yet support anything else besides EL &
CL and is not connected to the main parser.

Related to https://github.com/ethereum-optimism/optimism/issues/15152