refactor: Explicitlly specify rollup-boost socket addr in cli args (#271)
A recent audit outlined that the default socket address at which
rollup-boost establishes the RPC server is `0.0.0.0` which makes it
vulnerable to misconfiguration. I'm working on a PR to change the
default socket address to `127.0.0.1` but it currently breaks the
kurtosis integration tests.

This PR fixes it by explicitly setting the RPC host to `0.0.0.0`