feat(grafana): Grizzly compatibility layer (#280)
## Overview

Adds a grizzly compatibility layer in the grafana launcher. Grafana's
native dashboard exporting functionality is not compatible with grizzly,
and this makes it difficult to import dashboards.

Now, the grafana launcher will check if the dashboard is already
compatible with grizzly, and if not, will re-export the dashboard and
set Grafana's `DS_PROMETHEUS` variable to `Prometheus`.

### Periphery changes

- Imports `kona-node` and `op-reth` dashboards by default
- Removes the requirement for a dashboard source to define folders.