chore: Support multiple sequencers: DA [8/N] (#285)
**Description**

I resisted the urge to dive deeper into the DA configuration and kept
the existing logic as is, just moved it to the new format.

At the moment, we are accepting any combination of `image` and `cmd` so
DA can be truly anything, a static nginx server if one desires.

Related to https://github.com/ethereum-optimism/optimism/issues/15152