chore: Support multiple sequencers: Sequencers assignment [11/N] (#289)
**Description**

Introduces a new `sequencer` property of a network participant. Having
thought long and hard about the naming and purpose I decided this
property to be heterogenous, type-wise speaking.

The property accepts three kinds of values:

- `bool True` means that this node is a sequencer
- `bool False/None/falsy value` means that this node is **not** a
sequencer
- `string` means that this node is not a sequencer and the string refers
to a name of a node that is a sequencer and that this node will use when
starting the EL client (look for `--rollup.sequencerhttp` CLI flag)

The reason behind keeping it in one property instead of having a
property called `is_sequencer` and another one called e.g.
`sequencer-name` is the fact that two-property setup is redundant
(information-content-wise) and means that the pieces of information
stored in the two properties can be calculated based on one another
which also means they can drift apart.

If sequencers are not explicitly set, round-robin is used to select them
for the nodes. I am not set on this approach, we can go for something
else.