chore: Support multiple sequencers: No empty networks & sequencer selectors [12/N] (#292)
**Description**

We no longer allow network configurations with no participants. If the
network configuration is omited as a whole (i.e. it is `None`, then a
default network with one participant will be provided).

Set of small sequencer helpers is also introduced to encapsulate the
sequencer selection logic.

Edit: I also noticed that even though we mapped the `sequencer` property
to a string always, I forgot to undo the mapping to `True` in
`_apply_sequencers`. This has been fixed

Related to https://github.com/ethereum-optimism/optimism/issues/15152