chore: Support multiple sequencers: Move MEV under participant [13/N] (#294)
**Description**

MEV (rollup boost) is a per-sequencer service and needs to move under
the participant configuration if we are to support multiple sequencers.

This PR adjusts the new input parser and plumbs it back into the legacy
`package_io/input_parser.star` for the time being.

Related to https://github.com/ethereum-optimism/optimism/issues/15152