chore: Support multiple sequencers: Add EL URLs to parsed proxyd params [16/N] (#300)
**Description**

Since we know the EL URLs in advance, we can add them to the proxyd
parsed params instead of synthesizing them in the launcher. This makes
the launcher thinner and will make the transition from the old to the
new input parser a bit slimmer.

It requires a bit of forward plumbing in the current input parser - to
map the legacy parsed participants to the new format (only the bits
required by the `proxyd` launcher)

Related to https://github.com/ethereum-optimism/optimism/issues/15152