chore: Support multiple sequencers: New launchers for op-node and hildr [17/N] (#303)
**Description**

These are adjusted versions of the existing launchers for `op-node` and
`hildr`. They use the new input parser formats and have not been plugged
in yet.

The purpose of this PR is to reduce the size of the _great plugging in_
PR - these are not yet being used and the current ones have been left
untouched.

Related to https://github.com/ethereum-optimism/optimism/issues/15152