fix: Support multiple sequencers: Reorder the launch sequence [28/N] (#321)
**Description**

The launch sequence needs to be reorganized so that the conductors can
receive supervisor URLs (the supervisors need to be running by the time
we launch the conductors so we cannot simply pass the precomputed URLs).

All of this is heavily marked as `hack` since it's a dirty fix. It will
need to be replaced A$AP with a better solution - be it
https://github.com/ethpandaops/optimism-package/pull/318 or another.

Fixes https://github.com/ethereum-optimism/optimism/issues/16366