fix: Support multiple sequencers: Pass supervisors params to conductor [29/N] (#322)
**Description**

`op-conductor` needs to know about the supervisors, hence this.

We also need to pass the `rollup-boost` context to the conductor but
it's buried so deep in the `el_cl_launcher` that we'd need to go through
with the full refactor to support it (sad face).

Fixes https://github.com/ethereum-optimism/optimism/issues/16366