feat: op-signer: Add script for generating local credentials [4/N] (#376)
**Description**

This PR contains a line-by-line copy of [a script from the `op-signer`
repo](https://github.com/ethereum-optimism/infra/blob/main/op-signer/gen-local-creds.sh).
In this PR, the script is left untouched - the adjustments needed will
be included in the next PR to make it obvious what needs to be changed.

This script generates the necessary key and certificate files so that
the signer can communicate with the other services.

Related to
https://github.com/ethereum-optimism/platforms-team/issues/581