feat: op-signer: Launcher [6/N] (#380)
**Description**

In this PR the signer service is launched (if enabled in the args file).
It is not yet being connected to the batcher and proposer - that will
happen in the next PR.

The signer launch (or the configuration to be more specific) is not
straightforward - it requires a set of credential files and certificates
(generated by the shell scripts) as well as private key files (the HEX
private keys converted to PEM files).

For more pointers, follow:

- [The old, reverted
PR](https://github.com/ethpandaops/optimism-package/pull/207/files) that
changed too much, especially it reversed the way batcher, proposer &
challenger get their private keys and henceforth it got reverted
- [`op-signer`
README](https://github.com/ethereum-optimism/infra/tree/main/op-signer)

Related to
https://github.com/ethereum-optimism/platforms-team/issues/581